Add stateless validation for transactions
